+++ source/js - копия/admin.js
@@ -1,0 +1,3 @@
+var module = require('./module');
+
+console.log('admin run');
