+++ source/js - копия/map.js
@@ -1,199 +1,5 @@
-// index
-if($("div").is("#parallax")) {
-// parallax on index.htmt
-   (function(){
-     var parallaxContainer = document.getElementById('parallax'),
-        layers = parallaxContainer.children;
-
-var moveLayers = function (e) {
-   // console.log(e);
-    var initialX = (window.innerWidth / 2) - e.pageX,
-        initialY = (window.innerHeight / 2) - e.pageY;
-
-[].slice.call(layers).forEach(function (layer, i) {
-
-    var divider = i/100,
-        positionX = initialX * divider,
-        positionY = initialY * divider,
-        bottomPosition = (window.innerHeight / 2) * divider,
-        layerStyle = layer.style,
-        transformString = 'translate3d(' + positionX + 'px, ' + positionY + 'px, 0)';
-
- layerStyle.transform = transformString;
- layerStyle.bottom = '-' + bottomPosition + 'px';
-});
-}
-window.addEventListener('mousemove', moveLayers);
-})();
-// flip
-(function () {        
-$('#btn').on('click', function (e) {
-    e.preventDefault; 
-    $('#myCard').addClass('flip');
-    $(this).fadeOut(600);
-})
-})(); 
-} 
-
-
-
-
-// works
-else if ($("section").is(".works")) {
-// hamburger
+ // map
 (function(){
-  $('#toggle').click(function() {
-   $(this).toggleClass('active');
-   $('#overlay').toggleClass('open');
-  });
- })();
-
-// parallax по скроллу
-(function(){
- var parallax = function(){
-    var bg = document.querySelector('.hero__bg');
-    var user = document.querySelector('.user-block');
-    var sectionText = document.querySelector('.section__title');
-
-    return {
-        move: function (block, windowScroll, strafeAmount) {
-            var strafe = windowScroll / -strafeAmount + '%';
-            var transformString = 'translate3d(0, ' + strafe + ', 0)';
-            var style = block.style;
-
-            style.transform = transformString;
-            style.webkitTransform = transformString;
-
-        },
-    init: function (wScroll) {
-        this.move(bg, wScroll, 45);
-        this.move(sectionText, wScroll, 20);
-        this.move(user, wScroll, 3);
-        }
-    }
- }();
-
- window.onscroll = function () {
-    var wScroll = window.pageYOffset;
-    parallax.init(wScroll);
- }
- })();    
- 
-// blur
-(function() {
-$(document).ready(function(){
-  blur();
-})
-$(window).resize(function(){
-  blur();
-});
-
-function blur() {
-  var imgWidth = $('.blur__bg').width(),
-      blurSection = $('.blur'),
-      blur = $('.blur__form'),
-      posY = blurSection.offset().top - blur.offset().top,   //текущее положение элемента относительно документа.
-      posX = blurSection.offset().left - blur.offset().left;
-  
-  blur.css({
-      'background-size': imgWidth + 'px' + ' ' + 'auto',
-      'background-position': posX + 'px' + ' ' + posY + 'px'
-  })
-}
-})();
-}
-
-
-
-
-// blog
- else if ($("section").is(".blog")) {
-// hamburger
-(function(){
-  $('#toggle').click(function() {
-   $(this).toggleClass('active');
-   $('#overlay').toggleClass('open');
-  });
- })();
-
-// parallax по скроллу
-(function(){
- var parallax = function(){
-    var bg = document.querySelector('.hero__bg');
-    var user = document.querySelector('.user-block');
-    var sectionText = document.querySelector('.section__title');
-
-    return {
-        move: function (block, windowScroll, strafeAmount) {
-            var strafe = windowScroll / -strafeAmount + '%';
-            var transformString = 'translate3d(0, ' + strafe + ', 0)';
-            var style = block.style;
-
-            style.transform = transformString;
-            style.webkitTransform = transformString;
-
-        },
-    init: function (wScroll) {
-        this.move(bg, wScroll, 45);
-        this.move(sectionText, wScroll, 20);
-        this.move(user, wScroll, 3);
-        }
-    }
- }();
-
- window.onscroll = function () {
-    var wScroll = window.pageYOffset;
-    parallax.init(wScroll);
- }
- })();    
-} 
-
-
-
-
-
-// about
-else if ($("section").is(".about")) {
-// hamburger
-(function(){
-  $('#toggle').click(function() {
-   $(this).toggleClass('active');
-   $('#overlay').toggleClass('open');
-  });
- })();
-
-// parallax по скроллу
-(function(){
- var parallax = function(){
-    var bg = document.querySelector('.hero__bg');
-    var user = document.querySelector('.user-block');
-    var sectionText = document.querySelector('.section__title');
-
-    return {
-        move: function (block, windowScroll, strafeAmount) {
-            var strafe = windowScroll / -strafeAmount + '%';
-            var transformString = 'translate3d(0, ' + strafe + ', 0)';
-            var style = block.style;
-
-            style.transform = transformString;
-            style.webkitTransform = transformString;
-
-        },
-    init: function (wScroll) {
-        this.move(bg, wScroll, 45);
-        this.move(sectionText, wScroll, 20);
-        this.move(user, wScroll, 3);
-        }
-    }
- }();
-
- window.onscroll = function () {
-    var wScroll = window.pageYOffset;
-    parallax.init(wScroll);
- }
- })(); 
- //map
- (function(){
 google.maps.event.addDomListener(window, 'load', init);
     var map, markersArray = [];
 
@@ -355,34 +161,3 @@ google.maps.event.addDomListener(window, 'load', init);
     }
 })();
 
-
-} 
-
-
-// login
- //else ($("form").is("#order-form")) {
-//     (function(){
-//      var parallaxContainer = document.getElementById('parallax'),
-//         layers = parallaxContainer.children;
-
-// var moveLayers = function (e) {
-//    // console.log(e);
-//     var initialX = (window.innerWidth / 2) - e.pageX,
-//         initialY = (window.innerHeight / 2) - e.pageY;
-
-// [].slice.call(layers).forEach(function (layer, i) {
-
-//     var divider = i/100,
-//         positionX = initialX * divider,
-//         positionY = initialY * divider,
-//         bottomPosition = (window.innerHeight / 2) * divider,
-//         layerStyle = layer.style,
-//         transformString = 'translate3d(' + positionX + 'px, ' + positionY + 'px, 0)';
-
-//  layerStyle.transform = transformString;
-//  layerStyle.bottom = '-' + bottomPosition + 'px';
-// });
-// }
-// window.addEventListener('mousemove', moveLayers);
-// })();
-//};
