+++ source/js/flip.js
@@ -1,0 +1,10 @@
+// flip
+var flip = function () {        
+  
+$('#btn').on('click', function (e) {
+    e.preventDefault; 
+    $('#myCard').addClass('flip');
+    $(this).fadeOut(600);
+})
+};
+
