+++ source/js/hamburger.js
@@ -1,0 +1,11 @@
+// hamburger
+var hamburger = function(){
+  $('#toggle').click(function() {
+   $(this).toggleClass('active');
+   $('#overlay').toggleClass('open');
+  });
+ };
+
+
+
+
