+++ source/style/common/about.sass
@@ -6,7 +6,6 @@
 	
 .hero
 	background: url("../img/bg/portfolio.jpg") no-repeat center
-	//overflow: hidden
 	position: relative
 	height: 100vh
 	min-height: 650px
@@ -165,7 +164,6 @@
 
 .section__title 
 	display: inline-block
-	//@include centered(x)
 	@include laptop
 		display: none
