+++ source/style/common/blur.scss
@@ -2,7 +2,7 @@
 //   //height: 100%;
 //   position: relative;
 //   //overflow: hidden;
-//   //padding-top: 1px;
+//   
 // }
 
 // .blur__img {
