+++ source/style/common/welcome.sass
@@ -1,17 +1,8 @@
-//.welcome 
-	//background: url("../img/bg/portfolio.jpg") no-repeat center
-	//background-size: cover
-	//position: absolute
-	// top: 0
-	// left: 0
-	// right: 0
-	// bottom: 0
-
 .welcome__parallax
 	overflow: hidden
 	width: 100%
 	height: 100%
-	position: relative   //
+	position: relative   
 	
 .parallax
 	position: absolute
@@ -19,21 +10,18 @@
 	left: 0
 	width: 100%
 	height: 100%
-
+	
 .parallax__layer
-	height: 100%
 	position: absolute
-	width: 1800px
+	height: 100%
+	width: 2200px
 	left: 50%
-	margin-left: -900px
+	margin-left: -1100px
 	img
 		position: absolute
 		bottom: 0
-		max-width: 100%  //
-
-		
-		
-
+		max-width: 100%
+	
 .section__user-block
 	@include centered(both)
 
@@ -76,10 +64,15 @@
 	border-radius: 5px
 	overflow: hidden
 	border-radius: 5px
+	@include desktop
+		width: 350px
+		height: 388px
 	@include laptop
 		width: 350px
+		height: 380px
 	@include tablets
 		width: 300px
+		height: 368px
 
 .user-blok--transparent
 	background-color: transparent
@@ -163,10 +156,14 @@
 
 .form__elem
 	display: flex
-	margin: 0 20px 20px
+	margin: 0 20px 25px
 	border-radius: 5px 
 	overflow: hidden
 	background-color: #fff
+	@include desktop
+		margin-bottom: 20px
+	@include tablets
+		margin-bottom: 22px
 	&-checkbox
 		display: block
 		margin: 0 20px 10px
@@ -304,10 +301,6 @@
 .flip-container 
 	perspective: 1000px
 
-	/* flip the pane when hovered */
-// .flip-container .flipper, .flip-container.hover .flipper .flip-container.flip .flipper
-//.flip-container.hover .flipper, .flip-container.flip .flipper
-	//transform: rotateY(180deg)
 .flip 
 	.flipper
  		transform: rotateY(180deg)
@@ -322,7 +315,6 @@
 		width: 300px
 		height: 375px
 
-/* flip speed goes here */
 .flipper 
 	transition: 0.6s
 	transform-style: preserve-3d
@@ -336,23 +328,16 @@
 		width: 300px
 		height: 375px
 
-/* hide back of pane during swap */
 .front, .back 
 	backface-visibility: hidden
 	position: absolute
 	top: 0
 	left: 0
 
-/* front pane, placed above back */
 .front 
 	z-index: 2
-	/* for firefox 31 */
 	transform: rotateY(0deg)
 
-/* back, initially hidden pane */
 .back 
 	transform: rotateY(180deg)
 
-
-// .section__user-block
-// 	transform: skewY(10deg) skewY(-10deg)
