+++ source/template/pages/about - копия.pug
@@ -1,0 +1,167 @@
+extends ../_template
+include ../parts/mixins
+
+block variables
+  - var title = 'About me'
+    
+block content
+  section.section.hero
+    .hero__bg
+      img(src="assets/img/bg/portfolio.jpg", alt="").hero__bg-img
+
+    ul.socials.hero__socials
+      each item in ['vk', 'gh', 'in']
+        li.socials__item
+          a(href="#" class=`socials__link socials__link_${item}`)
+
+    .hero__hamburger
+    
+      .hamburger#toggle
+        .hamburger__line
+        .hamburger__line
+        .hamburger__line
+
+      .overlay#overlay
+        nav.overlay__menu
+          ul.overlay__menu-list
+            li.overlay__menu-item
+              a.overlay__menu-link(href="works.html") Мои работы
+            li.overlay__menu-item
+              a.overlay__menu-link(href="about.html") Обо мне
+            li.overlay__menu-item
+              a.overlay__menu-link(href="blog.html") Блог
+            li.overlay__menu-item
+              a.overlay__menu-link(href="login.html") Авторизация
+   
+    a(href="").down-arrow.hero__down-arrow
+      img(src="assets/img/icons/svg/arrow_down.svg", alt="").down-arrow__icon
+    
+    .hero__section
+      img(src="assets/img/decor/portfolio-header.png", alt="").section__title
+
+    .hero__user
+      .user-block.user-blok--transparent
+        .user__photo-wrap
+          .user__photo
+            img(src="assets/img/content/avatar-small.jpg", alt="").photo-avatar
+        .user__name.user__name--big Наталия Петкевич
+        .user__desc Личный сайт веб-разработчика
+      
+  section.section.about
+    .about__triangle
+      .about__triangle-left
+        svg(
+          viewbox="0 0 1000 100" 
+          preserveAspectRatio="none"
+          class="triangle triangle_fill_white about__triangles")
+          polygon(points="0,0 0,100 1000,100")
+      .about__triangle-right
+        svg(
+          viewbox="0 0 1000 100" 
+          preserveAspectRatio="none"
+          class="triangle triangle_fill_beige about__triangles")
+          polygon(points="0,100 1000,100 1000,0")
+    .about__left
+      .about__left-title
+        .about__section-title
+          .title.title--about
+            | Обо мне
+      .about__left-user
+        .user-about
+          .user-about__picture
+            img(src="assets/img/content/avatar.jpg", alt="").photo-avatar
+          .user-about__title
+           .title.title--subtitle
+              | Кто я
+          .user-about__text.text
+           p Я веб разработчик из Гродно. Мне 31 год. Я занимаюсь разработкой современных сайтов и приложений. Мне нравится делать интересные и современные проекты.
+           p Этот сайт я сделала в рамках обучения в Школе онлайн образования LoftSchool. Чуть позже я освежу в нем свой контент. А пока посмотрите, как тут все классно и красиво!
+    .about__right
+      .about-data
+        .about-data__description.about-user-info
+          .about-user-info__title
+            .title.title--subtitle
+              | Чем я могу быть полезен
+          .about-user-info__text.text
+            p Больше всего меня привлекает Frontend разработка, но я также знакома и могу решать не сложные задачи на Backend. Но давайте по порядку.
+        .about-data__skills
+            ul.skills 
+              li.skills__item
+                .skills__item-name Frontend
+                ul.skills-section__list
+                  each skillTitle in ['HTML5', 'CSS3', 'JavaScript&jQuery']
+                    li.skills-section__item
+                      .skill
+                        //svg(class="skill__circle `skill__circle--${skillTitle}`)
+                        .skill__name= skillTitle
+                .skills__item-name Backend
+                ul.skills-section__list
+                  each skillTitle in ['PHP', 'mySQL', 'Node.js  &npm', 'Mongo.db']
+                    li.skills-section__item
+                      .skill
+                        svg.skill__circle
+                        .skill__name= skillTitle
+                .skills__item-name WorkFlow
+                ul.skills-section__list
+                  each skillTitle in ['Git', 'Gulp', 'Bower']
+                    li.skills-section__item
+                      .skill
+                        svg.skill__circle
+                        .skill__name= skillTitle
+
+  section.section.map
+    .map__container#map
+
+    .contacts.map__contacts
+      .contacts__title.title.title--small
+        | Контакты
+      ul.contacts__list
+        li.contacts__item
+          a(href="#").contacts__link
+            img(src="assets/img/icons/svg/skype.svg").contacts__icon
+            .contacts__text Petkevich Natalia
+        li.contacts__item
+          a(href="#").contacts__link
+            img(src="assets/img/icons/svg/envelope.svg").contacts__icon
+            .contacts__text petkevich.natalli@gmail.com
+        li.contacts__item
+          a(href="#").contacts__link
+            img(src="assets/img/icons/svg/phone.svg").contacts__icon
+            .contacts__text +375 29 75 85 340
+        li.contacts__item
+          a(href="#").contacts__link
+            img(src="assets/img/icons/svg/map_marker.svg").contacts__icon
+            .contacts__text г.Гродно, ул. Захарова, д. 24
+
+block footer
+  footer.footer
+    .footer__line
+      .footer__left
+        ul.menu.footer__menu
+          li.menu__item
+            a(href="works.html").menu__link Мои работы
+          li.menu__item
+            a(href="about.html").menu__link Обо мне
+          li.menu__item
+            a(href="blog.html").menu__link Блог
+          li.menu__item
+            a(href="login.html").menu__link Авторизация
+      .footer__right
+        ul.socials.footer__socials
+          each item in ['vk', 'gh', 'in']
+            li.socials__item
+              a(href="#" class=`socials__link socials__link_${item}`)
+
+    .footer__line
+      .footer__left.footer__about-me
+        | Я веб разработчик из Гродно, и этот сайт я сделала в рамках обучения в школе онлайн образования Loftschool
+      .footer__right.footer__about-me
+        | &copy; Наталия Петкевич | создано с любовью в LoftSchool | 2017
+
+  
+  script(src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDwNILHWU2YjDjBQKy2SqVXwpkA5UoaHNs&extension=.js")
+  script(src="https://cdn.mapkit.io/v1/infobox.js")
+  link(href="https://fonts.googleapis.com/css?family=Roboto:300,400" rel="stylesheet") 
+  link(href="https://cdn.mapkit.io/v1/infobox.css" rel="stylesheet") 
+
+
