+++ source/template/pages/blog - копия.pug
@@ -1,0 +1,132 @@
+extends ../_template
+include ../parts/mixins
+
+block variables
+  - var title = 'Blog'
+    
+block content
+  section.section.hero
+    .hero__bg
+      img(src="assets/img/bg/portfolio.jpg", alt="").hero__bg-img
+
+    ul.socials.hero__socials
+      each item in ['vk', 'gh', 'in']
+        li.socials__item
+          a(href="#" class=`socials__link socials__link_${item}`)
+
+    .hero__hamburger
+    
+      .hamburger#toggle
+        .hamburger__line
+        .hamburger__line
+        .hamburger__line
+
+      .overlay#overlay
+        nav.overlay__menu
+          ul.overlay__menu-list
+            li.overlay__menu-item
+              a.overlay__menu-link(href="works.html") Мои работы
+            li.overlay__menu-item
+              a.overlay__menu-link(href="about.html") Обо мне
+            li.overlay__menu-item
+              a.overlay__menu-link(href="blog.html") Блог
+            li.overlay__menu-item
+              a.overlay__menu-link(href="login.html") Авторизация
+
+    //a(href="").down-arrow.hero__down-arrow
+      //img(src="assets/img/icons/svg/arrow_down.svg", alt="").down-arrow__icon
+
+    .hero__section
+      img(src="assets/img/decor/blog-header.png", alt="").section__title.section__title--blog 
+    
+    .hero__user
+      .user-block.user-blok--transparent
+        .user__photo-wrap
+          .user__photo
+            img(src="assets/img/content/avatar-small.jpg", alt="").photo-avatar
+        .blog__section-title 
+          .title.title--blog Блог
+        .user__desc Статьи, которые я написала
+
+  section.section.blog
+    .about__triangle
+      .about__triangle-left
+        svg(
+          viewbox="0 0 1000 100" 
+          preserveAspectRatio="none"
+          class="triangle triangle_fill_fff about__triangles")
+          polygon(points="0,0 0,100 1000,100")
+      .about__triangle-right
+        svg(
+          viewbox="0 0 1000 100" 
+          preserveAspectRatio="none"
+          class="triangle triangle_fill_fff about__triangles")
+          polygon(points="0,100 1000,100 1000,0")
+   
+    .blog__left
+      ul.sidebar-menu
+        li.sidebar-menu__item.item-active
+          a(href="").sidebar-menu__link
+            | Самое важное о SASS
+        li.sidebar-menu__item
+          a(href="").sidebar-menu__link
+            | Приемы в верстке, без которых не обходится не один сайт
+        li.sidebar-menu__item
+          a(href="").sidebar-menu__link
+            | Необходимый набор Gulp плагинов
+        li.sidebar-menu__item
+          a(href="").sidebar-menu__link
+            | Почему я выбрал Jade
+
+    .blog__right
+      .article
+        .article__title Самое важное в SASS
+        .article__data 22 ноября 2016
+        .article__content 
+          p.
+            Таким образом начало повседневной работы по формированию позиции позволяет выполнять важные задания по разработке направлений прогрессивного развития. Разнообразный и богатый опыт новая модель организационной деятельности играет важную роль в формировании новых предложений. Товарищи! новая модель организационной деятельности играет важную роль в формировании систем массового участия.
+          p.
+            Не следует, однако забывать, что укрепление и развитие структуры обеспечивает широкому кругу (специалистов) участие в формировании систем массового участия. Идейные соображения высшего порядка, а также рамки и место обучения кадров обеспечивает широкому кругу (специалистов) участие в формировании позиций, занимаемых участниками в отношении поставленных задач. Товарищи! постоянный количественный рост и сфера нашей активности способствует подготовки и реализации модели развития. С другой стороны рамки и место обучения кадров влечет за собой процесс внедрения и модернизации системы обучения кадров, соответствует насущным потребностям.
+        
+          img(src="assets/img/content/example.jpg" alt="")
+
+      .article
+        .article__title Приёмы в вёрстке, без которых не обходится ни один сайт
+        .article__data 13 ноября 2016
+        .article__content
+          p.
+            Таким образом начало повседневной работы по формированию позиции позволяет выполнять важные задания по разработке направлений прогрессивного развития. Разнообразный и богатый опыт новая модель организационной деятельности играет важную роль в формировании новых предложений. Товарищи! новая модель организационной деятельности играет важную роль в формировании систем массового участия.
+          p.
+            Не следует, однако забывать, что укрепление и развитие структуры обеспечивает широкому кругу (специалистов) участие в формировании систем массового участия. Идейные соображения высшего порядка, а также рамки и место обучения кадров обеспечивает широкому кругу (специалистов) участие в формировании позиций, занимаемых участниками в отношении поставленных задач. Товарищи! постоянный количественный рост и сфера нашей активности способствует подготовки и реализации модели развития. С другой стороны рамки и место обучения кадров влечет за собой процесс внедрения и модернизации системы обучения кадров, соответствует насущным потребностям.
+       
+block footer
+  footer.footer
+    .footer__line
+      .footer__left
+        ul.menu.footer__menu
+          li.menu__item
+            a(href="works.html").menu__link Мои работы
+          li.menu__item
+            a(href="about.html").menu__link Обо мне
+          li.menu__item
+            a(href="blog.html").menu__link Блог
+          li.menu__item
+            a(href="login.html").menu__link Авторизация
+      .footer__right
+        ul.socials.footer__socials
+          each item in ['vk', 'gh', 'in']
+            li.socials__item
+              a(href="#" class=`socials__link socials__link_${item}`)
+
+    .footer__line
+      .footer__left.footer__about-me
+        | Я веб разработчик из Гродно, и этот сайт я сделала в рамках обучения в школе онлайн образования Loftschool
+      .footer__right.footer__about-me
+        span
+          | &copy; Наталия Петкевич | создано с любовью в LoftSchool | 2017
+
+
+
+
+
+
