+++ source/template/pages/login - копия.pug
@@ -1,4 +1,4 @@
-extends ../_template-index
+extends ../_template
 include ../parts/mixins
 
 block variables
@@ -67,8 +67,6 @@ block content
           a(href="index.html").form__elem-main.form-menu__item На главную
                   
           input(type="submit" id="submit" name="submit" value="Войти").form__elem-submit.form-menu__item
-
-     
                   
   .footer__welcome
     | &copy; Наталия Петкевич | создано с любовью в LoftSchool | 2017
