+++ source/template/pages/works - копия.pug
@@ -1,0 +1,231 @@
+extends ../_template
+include ../parts/mixins
+
+block variables
+  - var title = 'Works'
+    
+block content
+  section.section.hero
+    .hero__bg
+      img(src="assets/img/bg/portfolio.jpg", alt="").hero__bg-img
+
+    ul.socials.hero__socials
+      each item in ['vk', 'gh', 'in']
+        li.socials__item
+          a(href="#" class=`socials__link socials__link_${item}`)
+
+    .hero__hamburger
+    
+      .hamburger#toggle
+        .hamburger__line
+        .hamburger__line
+        .hamburger__line
+
+      .overlay#overlay
+        nav.overlay__menu
+          ul.overlay__menu-list
+            li.overlay__menu-item
+              a.overlay__menu-link(href="works.html") Мои работы
+            li.overlay__menu-item
+              a.overlay__menu-link(href="about.html") Обо мне
+            li.overlay__menu-item
+              a.overlay__menu-link(href="blog.html") Блог
+            li.overlay__menu-item
+              a.overlay__menu-link(href="login.html") Авторизация
+
+
+    a(href="").down-arrow.hero__down-arrow
+      img(src="assets/img/icons/svg/arrow_down.svg", alt="").down-arrow__icon
+    
+    .hero__section
+      img(src="assets/img/decor/portfolio-header.png", alt="").section__title
+
+    .hero__user      
+      .user-block.user-blok--transparent
+        .user__photo-wrap
+          .user__photo
+            img(src="assets/img/content/avatar-small.jpg", alt="").photo-avatar
+        .user__name.user__name--big Наталия Петкевич
+        .user__desc Личный сайт веб-разработчика
+
+  section.section.works
+    .about__triangle
+      .about__triangle-left
+        svg(
+          viewbox="0 0 1000 100" 
+          preserveAspectRatio="none"
+          class="triangle triangle_fill_white about__triangles")
+          polygon(points="0,0 0,100 1000,100")
+      .about__triangle-right
+        svg(
+          viewbox="0 0 1000 100" 
+          preserveAspectRatio="none"
+          class="triangle triangle_fill_beige about__triangles")
+          polygon(points="0,100 1000,100 1000,0")
+
+    .works__title
+      .works__section-title
+        .title
+          | Мои работы
+
+    .slider.works__slider
+      .slider
+        .slider__controls
+            a(href="#").slider__btn.slider__btn--next
+              i.icon-chevron-right
+            a(href="#").slider__btn.slider__btn--prev
+              i.icon-chevron-left
+
+        .slider__nav
+          ul.slider__btn-list
+            li.slider__btn
+              a(href="").slider__btn-link
+            li.slider__btn
+              a(href="").slider__btn-link
+            li.slider__btn
+              a(href="").slider__btn-link
+            li.slider__btn
+              a(href="").slider__btn-link
+
+        .slider__wrap
+          ul.slider__list
+            li.slider__item.slider_item--active
+              .slider__photo
+                img(src="assets/img/slider/work-1.png", alt="").photo-slider
+              .slider__descr.descr
+                .descr__title Сайт школы онлайн Образования
+                .descr__technology HTML , CSS, JAVASCRIPT
+                .descr__site
+                  a(href="#").decsr__btn 
+                    i.icon-link
+                    span Посмотреть сайт
+            li.slider__item.slider_item--active
+              .slider__photo
+                img(src="assets/img/slider/work-2.png", alt="").photo-slider
+              .slider__descr.descr
+                .descr__title Сайт школы онлайн Образования
+                .descr__technology HTML , CSS, JAVASCRIPT
+                .descr__site
+                  a(href="#").decsr__btn 
+                    i.icon-link
+                    span Посмотреть сайт
+            li.slider__item.slider_item--active
+              .slider__photo
+                img(src="assets/img/slider/work-3.png", alt="").photo-slider
+              .slider__descr.descr
+                .descr__title Сайт школы онлайн Образования
+                .descr__technology HTML , CSS, JAVASCRIPT
+                .descr__site
+                  a(href="#").decsr__btn 
+                    i.icon-link
+                    span Посмотреть сайт
+            li.slider__item.slider_item--active
+              .slider__photo
+                img(src="assets/img/slider/work-4.png", alt="").photo-slider
+              .slider__descr.descr
+                .descr__title Сайт школы онлайн Образования
+                .descr__technology HTML , CSS, JAVASCRIPT
+                .descr__site
+                  a(href="#").decsr__btn 
+                    i.icon-link
+                    span Посмотреть сайт
+    
+
+  section.section
+    .blur
+      .reviews    
+        .reviews__section-title
+          .title Что обо мне говорят
+        .reviews__items
+          .review
+            i.icon-quote-left
+            .review__text.
+              Этот парень проходил обучение веб-разработке не где-то, а в LoftSchool! 2 месяца только самых тяжелых испытаний и бессонных ночей!
+            .reviews__autor
+              .reviews__autor-photo
+                img(src="assets/img/content/review-1.jpg", alt="").autor-photo
+              .reviews__autor-descr
+                .reviews__autor-name Ковальчук Дмитрий
+                .reviews__autor-position основатель Loftschool
+          .review
+            i.icon-quote-left
+            .review__text.
+              Этот код выдержит любые испытания. Только пожалуйста, не загружается сайт на слишком старых браузерах!
+            .reviews__autor
+              .reviews__autor-photo
+                img(src="assets/img/content/review-2.jpg", alt="").autor-photo
+              .reviews__autor-descr
+                .reviews__autor-name Сабанцев Владимир 
+                .reviews__autor-position преподаватель 
+
+
+
+
+      .blur__bg
+        img(src="assets/img/content/water.jpg", alt="").blur__img
+      
+         
+      .blur__form-wrapper
+        .blur__form
+
+        form(action="mail.php" method="POST" class="form-contact" id="contact-form")
+      
+          .title.title--subtitle
+            | Связаться со мной
+
+          label(class="form__elem" for="name")
+            input(class="form__elem-input" type="text" id="name" name="name" placeholder="Имя")
+
+          label(class="form__elem" for="email")
+            input(class="form__elem-input" type="text" id="email" name="email" placeholder="Email")
+
+          textarea(class="form__textarea form__elem" id="comment" name="comment" cols="42" rows="7" placeholder="Ваше сообщение")
+                     
+          .user__line-green
+            a(href="index.html").form__elem-main.form-menu__item На главную
+                  
+            input(type="reset" id="reset" name="reset" value="Очистить").form__elem-reset.form-menu__item
+
+      a(href="").down-arrow.reviews__down-arrow
+        img(src="assets/img/icons/svg/arrow_down.svg", alt="").down-arrow__icon
+
+            
+block footer
+  footer.footer
+    .footer__line
+      .footer__left
+       ul.menu.footer__menu
+          li.menu__item
+            a(href="works.html").menu__link Мои работы
+          li.menu__item
+            a(href="about.html").menu__link Обо мне
+          li.menu__item
+            a(href="blog.html").menu__link Блог
+          li.menu__item
+            a(href="login.html").menu__link Авторизация
+      .footer__right
+        ul.socials.footer__socials
+          each item in ['vk', 'gh', 'in']
+            li.socials__item
+              a(href="#" class=`socials__link socials__link_${item}`)
+
+    .footer__line
+      .footer__left.footer__about-me
+        | Я веб разработчик из Гродно, и этот сайт я сделала в рамках обучения в школе онлайн образования Loftschool
+      .footer__right.footer__about-me
+        | &copy; Наталия Петкевич | создано с любовью в LoftSchool | 2017
+
+
+
+
+      
+
+
+
+     
+
+  
+
+
+
+
